follow-the-line/requirements.txt
absl-py==0.15.0
apt-clone==0.2.1
apturl==0.5.2
asn1crypto==0.24.0
astunparse==1.6.3
beautifulsoup4==4.6.0
blinker==1.4
Brlapi==0.6.6
cachetools==4.2.4
certifi==2025.4.26
chardet==3.0.4
charset-normalizer==2.0.12
crccheck==1.3.0
cryptography==2.1.4
cupshelpers==1.0
cycler==0.10.0
Cython==0.29.21
dataclasses==0.8
decorator==4.1.2
defer==1.0.6
depthai==2.26.0.0
distlib==0.4.0
distro-info===0.18ubuntu0.18.04.1
evdev==1.6.1
feedparser==5.2.1
filelock==3.4.1
flatbuffers==25.2.10
gast==0.3.3
gdown==4.7.3
google-auth==2.22.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
graphsurgeon==0.4.5
grpcio==1.48.2
grpcio-tools==1.48.2
h5py==2.10.0
html5lib==0.999999999
httplib2==0.9.2
idna==3.10
importlib-metadata==4.8.3
importlib-resources==5.4.0
Jetson.GPIO==2.0.17
Keras-Applications==1.0.8
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
louis==3.5.0
lxml==4.2.1
macaroonbakery==1.1.3
Mako==1.0.7
Markdown==3.3.7
MarkupSafe==1.0
matplotlib==2.1.1
mock==5.2.0
numpy==1.19.5
oauth==1.0.1
oauthlib==3.2.2
onboard==1.4.1
opt-einsum==3.3.0
PAM==0.4.2
pandas==0.22.0
pbr==6.1.1
Pillow==8.4.0
platformdirs==2.4.0
protobuf==3.12.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pybind11==2.12.1
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
PyGObject==3.26.1
PyICU==1.9.8
PyJWT==1.5.3
pymacaroons==0.13.0
PyNaCl==1.1.2
pyparsing==2.2.0
pyRFC3339==1.0
pyserial==3.5
PySocks==1.7.1
python-apt==1.6.5+ubuntu0.7
python-dateutil==2.6.1
python-debian==0.1.32
python-version==0.0.2
pytz==2018.3
pyvesc @ git+https://github.com/LiamBindle/PyVESC@d2f2076afe91dafa920a38a437528819697e05e3
pyxattr==0.6.0
pyxdg==0.25
PyYAML==3.12
requests==2.27.1
requests-oauthlib==2.0.0
requests-unixsocket==0.1.5
rsa==4.9.1
scipy==0.19.1
SecretStorage==2.3.1
simplejson==3.13.2
six==1.15.0
ssh-import-id==5.7
system-service==0.3
systemd-python==234
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.4.0
tensorflow-estimator==2.5.0
tensorrt==8.2.1.8
termcolor==1.1.0
testresources==2.0.2
tqdm==4.64.1
typing-extensions==3.7.4.3
ubuntu-drivers-common==0.0.0
uff==0.6.9
unity-scope-calculator==0.1
unity-scope-chromiumbookmarks==0.1
unity-scope-colourlovers==0.1
unity-scope-devhelp==0.1
unity-scope-firefoxbookmarks==0.1
unity-scope-manpages==0.1
unity-scope-openclipart==0.1
unity-scope-texdoc==0.1
unity-scope-tomboy==0.1
unity-scope-virtualbox==0.1
unity-scope-yelp==0.1
unity-scope-zotero==0.1
UNKNOWN==0.0.0
urllib3==1.26.20
urwid==2.0.1
virtualenv==20.17.1
wadllib==1.3.2
webencodings==0.5
Werkzeug==2.0.3
wrapt==1.12.1
xkit==0.0.0
youtube_dl==2018.3.14
zipp==3.6.0
zope.interface==4.3.2
